+++ minimum-viable-sharing-meta.php
@@ -38,7 +38,7 @@ final class MinimumViableMeta_Core {
 	 * @since  1.0
 	 * @var    string
 	 */
-	private $version = '0.0.4';
+	private $version = '0.0.5';
 
 	/**
 	 * If an instance exists, this returns it.  If not, it creates one and
@@ -287,9 +287,8 @@ final class MinimumViableMeta_Core {
 			$query  = $wpdb->prepare("
 				SELECT post_id
 				FROM $table
-				WHERE meta_key = %s",
-				esc_sql( $metakey )
-			);
+				WHERE meta_key = %s
+			", esc_sql( $metakey ) );
 
 			// Run SQL query.
 			$theids = $wpdb->get_col( $query );
@@ -347,9 +346,8 @@ final class MinimumViableMeta_Core {
 		// Prepare my query.
 		$setup  = $wpdb->prepare("
 			DELETE FROM $table
-			WHERE meta_key = %s",
-			esc_sql( MINSHARE_META_POSTKEY )
-		);
+			WHERE meta_key = %s
+		", esc_sql( MINSHARE_META_POSTKEY ) );
 
 		// Run SQL query.
 		$query = $wpdb->query( $setup );
@@ -430,19 +428,23 @@ final class MinimumViableMeta_Core {
 		// Set my array of plugins and their keys.
 		$items  = array(
 			'yoast'     => array(
-				'name'  => __( 'Yoast SEO', 'minimum-viable-sharing-meta' ),
-				'title' => '_yoast_wpseo_title',
-				'desc'  => '_yoast_wpseo_metadesc',
+				'name'      => __( 'Yoast SEO', 'minimum-viable-sharing-meta' ),
+				'title'     => '_yoast_wpseo_title',
+				'desc'      => '_yoast_wpseo_metadesc',
+				'canonical' => '_yoast_wpseo_canonical',
+				'image'     => '_yoast_wpseo_opengraph-image',
 			),
 			'aioseo'    => array(
-				'name'  => __( 'All In One SEO Pack', 'minimum-viable-sharing-meta' ),
-				'title' => '_aioseop_title',
-				'desc'  => '_aioseop_description',
+				'name'      => __( 'All In One SEO Pack', 'minimum-viable-sharing-meta' ),
+				'title'     => '_aioseop_title',
+				'desc'      => '_aioseop_description',
+				'canonical' => '_aioseop_custom_link',
 			),
 			'genesis'   => array(
-				'name'  => __( 'Genesis Theme Framework', 'minimum-viable-sharing-meta' ),
-				'title' => '_genesis_title',
-				'desc'  => '_genesis_description',
+				'name'      => __( 'Genesis Theme Framework', 'minimum-viable-sharing-meta' ),
+				'title'     => '_genesis_title',
+				'desc'      => '_genesis_description',
+				'canonical' => '_genesis_canonical_uri',
 			),
 		);